Minimal support for wasm32-unknown-unknown
This introduces the --target-webasm-unknown option which switches to
building wasm32-unknown-unknown.

It's a crude patch as it only does the absolute minimum to be runnable
with a hand-crafted index.html, but I hope it helps discovering which
are the next steps.

One oddity that might be worth mentioning here is that as the compiler
already outputs .wasm as the only result, and cargo-web looks for an
additional s/.js/.wasm/ file, the WASM code is served both at /js/app.js
and its $CRATE.wasm file name -- this can be confusing when one tries to
use the default index.html, which loads /js/app.js as JavaScript and
fails at parsing it.