wasm-unknown-unknown: Serve adapted index.html
This serves the main WASM object at a static location (`/main.wasm`) and
serves a index.html suitable for running a JavaScript `eval` from within
Rust. (That mechanism is not intended for production use, but gets
things started until proper wrappers like from [1] are integrated).

Rust code that uses eval must include the following few lines for the time
being:

    extern "C" {
        fn js_raw_eval(pointer: *const u8, length: usize) -> ();
    }
    fn eval(expression: &str) {
        unsafe { js_raw_eval(expression.as_ptr(), expression.len()) }
    }

One caveat to be aware of is that the callbacks don't work while the
Rust `main()` function is executed, because the browser has not handed
out the instance object to JavaScript and it seems not to be available
directly in the callbacks. The current index.html exeutes a
`main_form_js` function instead (building as a library might be a better
idea on the long run), and passes around the instance as a global
variable (where a closure would be more suitable).

[1]: https://github.com/killercup/wasm-experiments