Support externally stored JS snippets
Storing the JS snippets as constants inside of the `.wasm`
is silly and unnecessarily bloats the module since we can't easily
get rid of them. Storing the JS snippets as custom sections is broken
due to Rust's custom section semantics being broken.

So the only thing we are left with is storing them as separate files
on the filesystem.